moved files into folders, merged AuthValidator and InputValidator, fixed all namespaces and removed unncessary dependencies
