+++ SmartSaver/Handler.cs
@@ -1,9 +1,8 @@
-﻿using DataManager;
-using System;
-using System.Collections.Generic;
+﻿using System;
 using System.Net.Http;
-using System.Text;
-using DataBases;
+using ePiggy.backend.calculations;
+using ePiggy.backend.dataHandling;
+using ePiggy.utilities;
 
 namespace ePiggy
 {
