+++ SmartSaver/Program.cs
@@ -1,16 +1,8 @@
 using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Linq;
 using System.Windows.Forms;
-using DataManager;
-using ePiggy;
 using ePiggy.forms;
-using ImageChartsLib;
-using System.Globalization;
-using DataBases;
 
-namespace EPiggy
+namespace ePiggy
 {
     static class Program
     {
