+++ SmartSaver/backend/calculations/OfferData.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace DataManager
+﻿namespace ePiggy.backend.calculations
 {
     public class OfferData
     {
