+++ SmartSaver/backend/dataHandling/Goal.cs
@@ -1,10 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Text;
 using System.Threading.Tasks;
 
-namespace DataManager
+namespace ePiggy.backend.dataHandling
 {
     public class Goal
     {
