+++ SmartSaver/backend/dataHandling/InternetParser.cs
@@ -1,17 +1,11 @@
-﻿using HtmlAgilityPack;
-using System;
-using System.Collections.Generic;
-using System.Diagnostics;
+﻿using System;
 using System.IO;
 using System.Linq;
 using System.Net;
-using System.Net.Http;
-using System.Net.Http.Headers;
-using System.Text;
 using System.Threading.Tasks;
-using ePiggy;
+using HtmlAgilityPack;
 
-namespace DataManager
+namespace ePiggy.backend.dataHandling
 {
     public static class InternetParser
     {
