+++ SmartSaver/db/Models/Goals.cs
@@ -1,9 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Text;
+﻿using System.ComponentModel.DataAnnotations;
 
-namespace DataBases
+namespace ePiggy.db.Models
 {
     public class Goals
     {
