+++ SmartSaver/db/Models/Incomes.cs
@@ -1,15 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using System.Configuration;
-using System.Data;
-using System.Text;
-using System.Windows.Forms;
-using Microsoft.EntityFrameworkCore;
 
-
-
-namespace DataBases
+namespace ePiggy.db.Models
 {
     public class Incomes
     {
