+++ SmartSaver/ePiggy.csproj
@@ -11,8 +11,8 @@
 
   <ItemGroup>
     <PackageReference Include="HtmlAgilityPack" Version="1.11.24" />
-    <Compile Remove="Migrations\20200927151042_InitialCreate.cs" />
-    <Compile Remove="Migrations\20200927151042_InitialCreate.Designer.cs" />
+    <Compile Remove="db\Migrations\20200927151042_InitialCreate.cs" />
+    <Compile Remove="db\Migrations\20200927151042_InitialCreate.Designer.cs" />
   </ItemGroup>
 
   <ItemGroup>
