+++ SmartSaver/forms/finances/budget/EntryInfoForm.cs
@@ -1,5 +1,5 @@
 ﻿using System.Windows.Forms;
-using DataManager;
+using ePiggy.backend.dataHandling;
 using ePiggy.utilities;
 
 namespace ePiggy.forms.finances.budget
