+++ SmartSaver/forms/finances/goals/GoalForm.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using System.Windows.Forms;
-using DataManager;
+using ePiggy.backend.dataHandling;
 using ePiggy.utilities;
 
 namespace ePiggy.forms.finances.goals
