+++ SmartSaver/forms/finances/goals/GoalsForm.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Windows.Forms;
-using DataManager;
+using ePiggy.backend.dataHandling;
 using ePiggy.utilities;
 
 namespace ePiggy.forms.finances.goals
