+++ SmartSaver/utilities/FormUtilities.cs
@@ -1,5 +1,4 @@
 ﻿using System.Windows.Forms;
-using ePiggy.forms.finances;
 using ePiggy.forms.finances.budget;
 using ePiggy.forms.finances.goals;
